Moved formatLocationForBackEnd() into utils.ts
